Find group of modules

Find group of modules based on shared genes

modGroups(modules, mset = NULL, min.overlap = 2, stat = "number")

Arguments

modules
Either a list of modules or character vector.

mset
Which module set to use. Either a character vector ("LI", "DC" or "all", default: LI) or an object of class tmod (see "Custom module definitions" below)

min.overlap
Minimum number of overlapping items if stat == number, minimum jaccard index if stat == jaccard etc.

stat
Type of statistics to return. "number": number of common genes (default); "jaccard": Jaccard index; "soerensen": Soerensen-Dice coefficient; "overlap": Szymkiewicz-Simpson coefficient.

Details

Split the modules into groups based on the overlapping items.

The first argument, modules, is either a character vector of module identifiers from `mset` (NULL mset indicates the default mset of tmod) or a list. If it is a list, then each element is assumed to be a character vector with module IDs.

Examples

mymods <- list(A=c(1, 2, 3), B=c(2, 3, 4), C=c(5, 6, 7))
modGroups(mymods)

#> $A
#> [1] "A" "B"
#> 
#> $C
#> [1] "C"
#>