Modules for metabolic profiling

Feature and data sets for metabolic profiling

Details

The module set "modmetabo" can be used with tmod to analyse metabolic profiling data. The clusters defined in this set are based on hierarchical clustering of metabolic compounds from human serum and have been published in a paper on metabolic profiling in tuberculosis by Weiner et al. (2012).

For an example analysis, "tbmprof" is a data set containing metabolic profiles of serum isolated from tuberculosis (TB) patients and healthy individuals. The tbmprof is a data frame containing observations in rows and metabolite id's (corresponding to the id's in the modmetabo object). See examples below.

References

Weiner et al. "Biomarkers of inflammation, immunosuppression and stress with active disease are revealed by metabolomic profiling of tuberculosis patients." PloS one 7.7 (2012): e40221.

See also

Examples

data(modmetabo)  # module definitions
data(tbmprof)    # example data set
ids <- rownames(tbmprof)
tb  <- factor(gsub("\\..*", "", ids))

## use Wilcoxon test to calculate significant differences
wcx <- apply(tbmprof, 2, function(x) wilcox.test(x ~ tb)$p.value)
wcx <- sort(wcx)
tmodCERNOtest(names(wcx), mset=modmetabo)
#>            ID                                           Title     cerno N1
#> ME.107 ME.107                             Amino acids cluster 104.64470 18
#> ME.37   ME.37 Kynurenines, taurocholates and cortisol cluster 116.88147 25
#> MP.2     MP.2                                      Amino Acid  99.16212 28
#>              AUC      cES      P.Value    adj.P.Val
#> ME.107 0.8824576 2.906797 1.283792e-08 5.391925e-07
#> ME.37  0.8837093 2.337629 2.816098e-07 5.913805e-06
#> MP.2   0.7060786 1.770752 3.357995e-04 4.701193e-03