Default gene expression module data

Gene expression module data from Chaussabel et al. (2008) and Li et al. (2014)

Details

The tmod package includes one data set of class tmod which can be loaded with data(tmod). This data set is derived from two studies (see package vignette for details). By default, enrichment analysis with tmod uses this data set; however, it is not loaded into user workspace by default.

References

Chaussabel, Damien, Charles Quinn, Jing Shen, Pinakeen Patel, Casey Glaser, Nicole Baldwin, Dorothee Stichweh, et al. 2008. "A Modular Analysis Framework for Blood Genomics Studies: Application to Systemic Lupus Erythematosus." Immunity 29(1):150-64.

Li, Shuzhao, Nadine Rouphael, Sai Duraisingham, Sandra Romero-Steiner, Scott Presnell, Carl Davis, Daniel S Schmidt, et al. 2014. "Molecular Signatures of Antibody Responses Derived from a Systems Biology Study of Five Human Vaccines." Nature Immunology 15(2):195-204.

See also

Examples

# list of first 10 modules
data(tmod)
tmod

#> An object of class "tmod"
#> 	606 modules, 12712 genes

tmod$MODULES[1:10, ]

#>              ID                                   Title           Category
#> LI.M0     LI.M0                      targets of FOSL1/2         TF targets
#> LI.M1.0 LI.M1.0  integrin cell surface interactions (I) molecular function
#> LI.M1.1 LI.M1.1 integrin cell surface interactions (II) molecular function
#> LI.M2.0 LI.M2.0                extracellular matrix (I)           location
#> LI.M2.1 LI.M2.1               extracellular matrix (II)           location
#> LI.M2.2 LI.M2.2              extracellular matrix (III)           location
#> LI.M3     LI.M3       regulation of signal transduction          signaling
#> LI.M4.0 LI.M4.0            cell cycle and transcription biological process
#> LI.M4.1 LI.M4.1                          cell cycle (I) biological process
#> LI.M4.2 LI.M4.2                   PLK1 signaling events          signaling
#>         Annotated
#> LI.M0         Yes
#> LI.M1.0       Yes
#> LI.M1.1       Yes
#> LI.M2.0       Yes
#> LI.M2.1       Yes
#> LI.M2.2       Yes
#> LI.M3         Yes
#> LI.M4.0       Yes
#> LI.M4.1       Yes
#> LI.M4.2       Yes
#>                                                                               URL
#> LI.M0     http://www.interactivefigures.com/meni/btm416_annotation/btmdata/M0.htm
#> LI.M1.0 http://www.interactivefigures.com/meni/btm416_annotation/btmdata/M1.0.htm
#> LI.M1.1 http://www.interactivefigures.com/meni/btm416_annotation/btmdata/M1.1.htm
#> LI.M2.0 http://www.interactivefigures.com/meni/btm416_annotation/btmdata/M2.0.htm
#> LI.M2.1 http://www.interactivefigures.com/meni/btm416_annotation/btmdata/M2.1.htm
#> LI.M2.2 http://www.interactivefigures.com/meni/btm416_annotation/btmdata/M2.2.htm
#> LI.M3     http://www.interactivefigures.com/meni/btm416_annotation/btmdata/M3.htm
#> LI.M4.0 http://www.interactivefigures.com/meni/btm416_annotation/btmdata/M4.0.htm
#> LI.M4.1 http://www.interactivefigures.com/meni/btm416_annotation/btmdata/M4.1.htm
#> LI.M4.2 http://www.interactivefigures.com/meni/btm416_annotation/btmdata/M4.2.htm
#>                 Source SourceID original.ID   B
#> LI.M0   Li et al. 2013       LI          M0  12
#> LI.M1.0 Li et al. 2013       LI        M1.0  29
#> LI.M1.1 Li et al. 2013       LI        M1.1  12
#> LI.M2.0 Li et al. 2013       LI        M2.0  30
#> LI.M2.1 Li et al. 2013       LI        M2.1  45
#> LI.M2.2 Li et al. 2013       LI        M2.2  14
#> LI.M3   Li et al. 2013       LI          M3  47
#> LI.M4.0 Li et al. 2013       LI        M4.0 335
#> LI.M4.1 Li et al. 2013       LI        M4.1 145
#> LI.M4.2 Li et al. 2013       LI        M4.2  34

tmod[1:10]

#> An object of class "tmod"
#> 	10 modules, 455 genes