Calculate AUC

Calculate AUC

tmodAUC(
  l,
  ranks,
  modules = NULL,
  stat = "AUC",
  recalculate.ranks = TRUE,
  filter = FALSE,
  mset = "LI",
  nodups = TRUE
)

Arguments

l
List of gene names corresponding to rows from the ranks matrix

ranks
a matrix with ranks, where columns correspond to samples and rows to genes from the l list

modules
optional list of modules for which to make the test

stat
Which statistics to generate. Default: AUC

recalculate.ranks
Filtering and removing duplicates will also remove ranks, so that they should be recalculated. Use FALSE if you don't want this behavior. If unsure, stay with TRUE

filter
Remove gene names which have no module assignments

mset
Which module set to use. "LI", "DC" or "all" (default: LI)

nodups
Remove duplicate gene names in l and corresponding rows from ranks

Value

A matrix with the same number of columns as "ranks" and as many rows as there were modules to be tested.

Details

tmodAUC calculates the AUC and U statistics. The main purpose of this function is the use in randomization tests. While tmodCERNOtest and tmodUtest both calculate, for each module, the enrichment in a single sorted list of genes, tmodAUC takes any number of such sorted lists. Or, actually, sortings -- vectors with ranks of the genes in each replicate. Note that the input for this function is different from tmodUtest and related functions: the ordering of l and the matrix ranks does not matter, as long as the matrix ranks contains the actual rankings. Each column in the ranks matrix is treated as a separate sample.

See also

Examples

data(tmod)
l <- tmod$GENES$ID
ranks <- 1:length(l)
res <- tmodAUC(l, ranks)
head(res)

#>              [,1]
#> LI.M0   1.0000000
#> LI.M1.0 0.9991463
#> LI.M1.1 0.9981890
#> LI.M2.0 0.9982022
#> LI.M2.1 0.9969878
#> LI.M2.2 0.9974349