Count the Up- or Down-regulated genes per module

For each module in a set, calculate the number of genes which are in that module and which are significantly up- or down-regulated.

tmodDecideTests(
  g,
  lfc = NULL,
  pval = NULL,
  lfc.thr = 0.5,
  pval.thr = 0.05,
  labels = NULL,
  filter.unknown = FALSE,
  mset = "all"
)

Arguments

g
a character vector with gene symbols

lfc
a numeric vector or a matrix with log fold changes

pval
a numeric vector or a matrix with p-values. Must have the same dimensions as lfc

lfc.thr
log fold change threshold

pval.thr
p-value threshold

labels
Names of the comparisons. Either NULL or a character vector of length equal to the number of columns in lfc and pval.

filter.unknown
If TRUE, modules with no annotation will be omitted

mset
Which module set to use. Either a character vector ("LI", "DC" or "all", default: LI) or a list (see "Custom module definitions" below)

Value

A list with as many elements as there were comparisons (columns in lfc and pval). Each element of the list is a data frame with the columns "Down", "Zero" and "Up" giving the number of the down-, not- and up-regulated genes respectively. Rows of the data frame correspond to module IDs.

Details

This function can be used to decide whether a module, as a whole, is up- or down regulated. For each module, it calculates the number of genes which are up-, down- or not regulated. A gene is considered to be up- regulated if the associated p-value is smaller than pval.thr and the associated log fold change is greater than lfc.thr. A gene is considered to be down- regulated if the associated p-value is smaller than pval.thr and the associated log fold change is smaller than lfc.thr.

Note that unlike decideTests from limma, tmodDecideTests does not correct the p-values for multiple testing -- therefore, the p-values should already be corrected.

See also