tmod's replacement for the limma topTable function

Produce a data frame for all or for selected coefficients of a linear fit object, including log fold changes, q-values, confidence intervals and MSD.

tmodLimmaTopTable(
  f,
  genelist = NULL,
  coef = NULL,
  adjust.method = "BH",
  confint = 0.95
)

Arguments

f
result of linear model fit produced by limma functions lmFit and eBayes

genelist
A data frame or a character vector with additional information on the genes / probes

coef
Which coefficients to extract

adjust.method
Which method of p-value adjustment; see "p.adjust()"

confint
Confidence interval to be calculated

Value

A data frame with all genes.

Details

Produce a data frame for all or for selected coefficients of a linear fit object, including log fold changes, q-values, confidence intervals and MSD. For each coefficient, these four columns will be created in the output file, with the name consisting of a prefix indicating the type of the column ("msd", "logFC", "qval", "SE", "ci.L", "ci.R") and the name of the coefficient.

See also