Modules for metabolic profiling

Feature and data sets for metabolic profiling

Details

The module set "modmetabo" can be used with tmod to analyse metabolic profiling data. The clusters defined in this set are based on hierarchical clustering of metabolic compounds from human serum and have been published in a paper on metabolic profiling in tuberculosis by Weiner et al. (2012).

For an example analysis, "tbmprof" is a data set containing metabolic profiles of serum isolated from tuberculosis (TB) patients and healthy individuals. The tbmprof is a data frame containing observations in rows and metabolite id's (corresponding to the id's in the modmetabo object). See examples below.

References

Weiner et al. "Biomarkers of inflammation, immunosuppression and stress with active disease are revealed by metabolomic profiling of tuberculosis patients." PloS one 7.7 (2012): e40221.

See also

Examples

data(modmetabo)  # module definitions
data(tbmprof)    # example data set
ids <- rownames(tbmprof)
tb  <- factor(gsub("\\..*", "", ids))

## use Wilcoxon test to calculate significant differences
wcx <- apply(tbmprof, 2, function(x) wilcox.test(x ~ tb)$p.value)
wcx <- sort(wcx)
tmodCERNOtest(names(wcx), mset=modmetabo)

#> # tmod report (class tmodReport) 8 x 3:
#>       │ID    │Title                                          │cerno│N1   │AUC  
#> ME.107│ME.107│Amino acids cluster                            │  105│   18│ 0.88
#>  ME.37│ ME.37│Kynurenines, taurocholates and cortisol cluster│  117│   25│ 0.88
#>   MP.2│  MP.2│Amino Acid                                     │   99│   28│ 0.71
#>       │cES  │P.Value│adj.P.Val
#> ME.107│  2.9│1.3e-08│5.4e-07  
#>  ME.37│  2.3│2.8e-07│5.9e-06  
#>   MP.2│  1.8│0.00034│0.0047