Create a summary of multiple tmod analyses

Create a summary of multiple tmod analyses

tmodSummary(
  x,
  clust = NULL,
  filter.empty = FALSE,
  filter.unknown = TRUE,
  select = NULL,
  effect.col = "AUC",
  pval.col = "adj.P.Val"
)

Arguments

x
list, in which each element has been generated with a tmod test function

clust
whether, in the resulting data frame, the modules should be ordered by clustering them with either q-values ("qval") or the effect size ("effect"). If "sort" or NULL, the modules are sorted alphabetically by their ID. If "keep", then the order of the modules is kept.

filter.empty
If TRUE, all elements (columns) with no significant enrichment will be removed

filter.unknown
If TRUE, modules with no annotation will be omitted

select
a character vector of module IDs to show. If clust == "keep", then in that particular order.

effect.col
The name of the column with the effect size

pval.col
The name of the p-value column

Value

a data frame with a line for each module encountered anywhere in the list x, two columns describing the module (ID and module title), and two columns(effect size and q value) for each element of list x.

Details

This function is useful if you run an analysis for several conditions or time points and would like to summarize the information in a single data frame. You can use lapply() to generate a list with tmod results and use tmodSummary to convert it to a data frame.

See also

Examples

data(Egambia)
E <- Egambia[,-c(1:3)]
pca <- prcomp(t(E), scale.=TRUE)

# Calculate enrichment for each component
gs   <- Egambia$GENE_SYMBOL
gn.f <- function(r) {
    tmodCERNOtest(gs[order(abs(r),
                decreasing=TRUE)],
                qval=0.01)
}
x <- apply(pca$rotation, 2, gn.f)
tmodSummary(x)

#> A tmodSummary object
#>                  ID                                              Title AUC.PC1
#> LI.M11.0   LI.M11.0                         enriched in monocytes (II)      NA
#> LI.M112.0 LI.M112.0                          complement activation (I)      NA
#> LI.M118.0 LI.M118.0                         enriched in monocytes (IV)      NA
#> LI.M127     LI.M127                         type I interferon response      NA
#> LI.M144     LI.M144                            cell cycle, ATP binding      NA
#> LI.M150     LI.M150                          innate antiviral response      NA
#> LI.M16       LI.M16                     TLR and inflammatory signaling      NA
#> LI.M165     LI.M165         enriched in activated dendritic cells (II)      NA
#> LI.M196     LI.M196                platelet activation - actin binding      NA
#> LI.M2.1     LI.M2.1                          extracellular matrix (II)      NA
#> LI.M213     LI.M213 regulation of transcription, transcription factors      NA
#> LI.M30       LI.M30      cell movement, Adhesion & Platelet activation      NA
#> LI.M37.0   LI.M37.0                immune activation - generic cluster      NA
#> LI.M37.1   LI.M37.1                        enriched in neutrophils (I)      NA
#> LI.M40       LI.M40              complement and other receptors in DCs      NA
#> LI.M47.0   LI.M47.0                            enriched in B cells (I)      NA
#> LI.M67       LI.M67                          activated dendritic cells      NA
#> LI.M75       LI.M75                            antiviral IFN signature      NA
#> LI.M81       LI.M81            enriched in myeloid cells and monocytes      NA
#> LI.S4         LI.S4                         Monocyte surface signature      NA
#> LI.S5         LI.S5                               DC surface signature      NA
#>           q.PC1 AUC.PC2 q.PC2   AUC.PC3        q.PC3   AUC.PC4        q.PC4
#> LI.M11.0     NA      NA    NA        NA           NA 0.7734395 2.136524e-07
#> LI.M112.0    NA      NA    NA        NA           NA 0.7509865 4.910746e-05
#> LI.M118.0    NA      NA    NA        NA           NA 0.8528591 5.027869e-05
#> LI.M127      NA      NA    NA        NA           NA 0.9593030 3.706095e-03
#> LI.M144      NA      NA    NA 0.9894257 0.0060518476        NA           NA
#> LI.M150      NA      NA    NA        NA           NA 0.9644887 2.780742e-03
#> LI.M16       NA      NA    NA        NA           NA 0.9923252 1.761180e-05
#> LI.M165      NA      NA    NA        NA           NA 0.7562436 1.423828e-05
#> LI.M196      NA      NA    NA        NA           NA 0.8375720 9.572552e-03
#> LI.M2.1      NA      NA    NA        NA           NA        NA           NA
#> LI.M213      NA      NA    NA 0.9610996 0.0002933835        NA           NA
#> LI.M30       NA      NA    NA        NA           NA        NA           NA
#> LI.M37.0     NA      NA    NA        NA           NA 0.7188785 2.335488e-19
#> LI.M37.1     NA      NA    NA        NA           NA 0.8671929 1.547673e-05
#> LI.M40       NA      NA    NA        NA           NA 0.7935087 3.395701e-03
#> LI.M47.0     NA      NA    NA        NA           NA 0.8564113 3.706902e-03
#> LI.M67       NA      NA    NA        NA           NA 0.9782503 4.910746e-05
#> LI.M75       NA      NA    NA        NA           NA 0.9007267 2.565179e-05
#> LI.M81       NA      NA    NA        NA           NA 0.7524813 3.959048e-03
#> LI.S4        NA      NA    NA        NA           NA 0.8090947 7.771600e-05
#> LI.S5        NA      NA    NA        NA           NA 0.7058762 5.837620e-04
#>           AUC.PC5 q.PC5 AUC.PC6 q.PC6 AUC.PC7 q.PC7 AUC.PC8 q.PC8   AUC.PC9
#> LI.M11.0       NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M112.0      NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M118.0      NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M127        NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M144        NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M150        NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M16         NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M165        NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M196        NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M2.1        NA    NA      NA    NA      NA    NA      NA    NA 0.6937974
#> LI.M213        NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M30         NA    NA      NA    NA      NA    NA      NA    NA 0.8690748
#> LI.M37.0       NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M37.1       NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M40         NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M47.0       NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M67         NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M75         NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.M81         NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.S4          NA    NA      NA    NA      NA    NA      NA    NA        NA
#> LI.S5          NA    NA      NA    NA      NA    NA      NA    NA        NA
#>                 q.PC9 AUC.PC10 q.PC10 AUC.PC11 q.PC11 AUC.PC12 q.PC12 AUC.PC13
#> LI.M11.0           NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M112.0          NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M118.0          NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M127            NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M144            NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M150            NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M16             NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M165            NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M196            NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M2.1   0.004728079       NA     NA       NA     NA       NA     NA       NA
#> LI.M213            NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M30    0.004728079       NA     NA       NA     NA       NA     NA       NA
#> LI.M37.0           NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M37.1           NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M40             NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M47.0           NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M67             NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M75             NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M81             NA       NA     NA       NA     NA       NA     NA       NA
#> LI.S4              NA       NA     NA       NA     NA       NA     NA       NA
#> LI.S5              NA       NA     NA       NA     NA       NA     NA       NA
#>           q.PC13  AUC.PC14      q.PC14 AUC.PC15 q.PC15 AUC.PC16 q.PC16 AUC.PC17
#> LI.M11.0      NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M112.0     NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M118.0     NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M127       NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M144       NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M150       NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M16        NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M165       NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M196       NA 0.8731049 0.003807907       NA     NA       NA     NA       NA
#> LI.M2.1       NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M213       NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M30        NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M37.0      NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M37.1      NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M40        NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M47.0      NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M67        NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M75        NA        NA          NA       NA     NA       NA     NA       NA
#> LI.M81        NA 0.7574997 0.007071980       NA     NA       NA     NA       NA
#> LI.S4         NA        NA          NA       NA     NA       NA     NA       NA
#> LI.S5         NA        NA          NA       NA     NA       NA     NA       NA
#>           q.PC17 AUC.PC18 q.PC18 AUC.PC19 q.PC19 AUC.PC20 q.PC20 AUC.PC21
#> LI.M11.0      NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M112.0     NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M118.0     NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M127       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M144       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M150       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M16        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M165       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M196       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M2.1       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M213       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M30        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M37.0      NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M37.1      NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M40        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M47.0      NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M67        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M75        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M81        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.S4         NA       NA     NA       NA     NA       NA     NA       NA
#> LI.S5         NA       NA     NA       NA     NA       NA     NA       NA
#>           q.PC21 AUC.PC22 q.PC22 AUC.PC23 q.PC23 AUC.PC24 q.PC24 AUC.PC25
#> LI.M11.0      NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M112.0     NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M118.0     NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M127       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M144       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M150       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M16        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M165       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M196       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M2.1       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M213       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M30        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M37.0      NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M37.1      NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M40        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M47.0      NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M67        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M75        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M81        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.S4         NA       NA     NA       NA     NA       NA     NA       NA
#> LI.S5         NA       NA     NA       NA     NA       NA     NA       NA
#>           q.PC25 AUC.PC26 q.PC26 AUC.PC27 q.PC27 AUC.PC28 q.PC28 AUC.PC29
#> LI.M11.0      NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M112.0     NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M118.0     NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M127       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M144       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M150       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M16        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M165       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M196       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M2.1       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M213       NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M30        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M37.0      NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M37.1      NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M40        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M47.0      NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M67        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M75        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.M81        NA       NA     NA       NA     NA       NA     NA       NA
#> LI.S4         NA       NA     NA       NA     NA       NA     NA       NA
#> LI.S5         NA       NA     NA       NA     NA       NA     NA       NA
#>           q.PC29  AUC.PC30       q.PC30
#> LI.M11.0      NA        NA           NA
#> LI.M112.0     NA        NA           NA
#> LI.M118.0     NA        NA           NA
#> LI.M127       NA        NA           NA
#> LI.M144       NA        NA           NA
#> LI.M150       NA        NA           NA
#> LI.M16        NA        NA           NA
#> LI.M165       NA        NA           NA
#> LI.M196       NA        NA           NA
#> LI.M2.1       NA        NA           NA
#> LI.M213       NA        NA           NA
#> LI.M30        NA        NA           NA
#> LI.M37.0      NA 0.6546699 3.097489e-05
#> LI.M37.1      NA 0.8905622 3.728092e-05
#> LI.M40        NA        NA           NA
#> LI.M47.0      NA        NA           NA
#> LI.M67        NA        NA           NA
#> LI.M75        NA        NA           NA
#> LI.M81        NA        NA           NA
#> LI.S4         NA        NA           NA
#> LI.S5         NA        NA           NA